LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-events-expected.txt


FAIL View timelime generates animationstart and animationend events assert_equals: Target initially off-screen and no animationstart event expected 0 but got 1
